SPA-142 Modify flags for MSVC to compile with all flag

+++ Team03/Code03/src/unit_testing/src/spa/sp/test_simple_parser.cpp
@@ -12,7 +12,7 @@ bool deepEqual(TNode node1, TNode node2) {
   if (node1.GetChildren().size() != node2.GetChildren().size()) {
     return false;
   }
-  for (int i = 0; i < node1.GetChildren().size(); i++) {
+  for (int i = 0; i < (int)node1.GetChildren().size(); i++) {
     if (!deepEqual(*node1.GetChildren()[i], *node2.GetChildren()[i])) {
       return false;
     }
